Rename fields and methods
