+++ src/test/java/com/mithwick93/refactoring/java/repositroy/impl/MovieRepositoryImplTest.java
@@ -1,7 +1,7 @@
-package com.mithwick93.refactoring.java.repositroy;
+package com.mithwick93.refactoring.java.repositroy.impl;
 
 import com.mithwick93.refactoring.java.entity.Movie;
-import com.mithwick93.refactoring.java.repositroy.impl.MovieRepositoryImpl;
+import com.mithwick93.refactoring.java.repositroy.MovieRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-class MovieRepositoryTest {
+class MovieRepositoryImplTest {
 
     private MovieRepository movieRepository;
