+++ src/test/java/com/mithwick93/refactoring/java/service/impl/RentalInformationServiceImplTest.java
@@ -1,9 +1,10 @@
-package com.mithwick93.refactoring.java.service;
+package com.mithwick93.refactoring.java.service.impl;
 
 import com.mithwick93.refactoring.java.entity.Customer;
 import com.mithwick93.refactoring.java.entity.MovieRental;
 import com.mithwick93.refactoring.java.repositroy.MovieRepository;
 import com.mithwick93.refactoring.java.repositroy.impl.MovieRepositoryImpl;
+import com.mithwick93.refactoring.java.service.RentalInformationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -13,20 +14,20 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class RentalInformationServiceTest {
+public class RentalInformationServiceImplTest {
     private RentalInformationService rentalInformationService;
 
     @BeforeEach
     void setUp() {
         MovieRepository movieRepository = new MovieRepositoryImpl();
-        rentalInformationService = new RentalInformationService(movieRepository);
+        rentalInformationService = new RentalInformationServiceImpl(movieRepository);
     }
 
     @Test
     void givenNullCustomer_whenStatement_thenThrowIllegalArgumentException() {
         Customer customer = null;
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -34,7 +35,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = new ArrayList<>();
         Customer customer = new Customer("", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -42,7 +43,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = new ArrayList<>();
         Customer customer = new Customer("C. U. Stomer", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -50,7 +51,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = List.of(new MovieRental(null, 3));
         Customer customer = new Customer("C. U. Stomer", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -58,7 +59,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = List.of(new MovieRental("F005", 3));
         Customer customer = new Customer("C. U. Stomer", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -66,7 +67,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = List.of(new MovieRental("F001", 0));
         Customer customer = new Customer("C. U. Stomer", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -74,7 +75,7 @@ public class RentalInformationServiceTest {
         List<MovieRental> movieRentals = List.of(new MovieRental("F001", -4));
         Customer customer = new Customer("C. U. Stomer", movieRentals);
 
-        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.statement(customer));
+        assertThrows(IllegalArgumentException.class, () -> rentalInformationService.getStatement(customer));
     }
 
     @Test
@@ -93,7 +94,7 @@ public class RentalInformationServiceTest {
                 You earned 2 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -110,7 +111,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -127,7 +128,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -144,7 +145,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -161,7 +162,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -178,7 +179,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -195,7 +196,7 @@ public class RentalInformationServiceTest {
                 You earned 2 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -212,7 +213,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -229,7 +230,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -246,7 +247,7 @@ public class RentalInformationServiceTest {
                 You earned 1 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
@@ -270,7 +271,7 @@ public class RentalInformationServiceTest {
                 You earned 4 frequent points
                 """;
 
-        String result = rentalInformationService.statement(customer);
+        String result = rentalInformationService.getStatement(customer);
 
         assertEquals(expected, result);
     }
